Attempt to fix the bootstrap css HTTP_404 error when deployed to Heroku
Found a variable called extra_css and extra_js in base.html template of
flask admin that can be used to add more css as a list to the template.

Thus created a function in IndexView and ModelView to get the css from
the controller thus enabling us to modify the css individually if nessary

But, this approach lead to some repeating code blocks.

TODO: make use of DRY for above implementation